+++ .vscode/settings.json
@@ -20,5 +20,14 @@
         "emscripten",
         "Relu"
     ],
-    "todo-tree.tree.showCountsInTree": true
+    "todo-tree.tree.showCountsInTree": true,
+
+    // The path to the `bun` executable.
+    "bun.runtime": "/home/jonas/.bun/bin/bun",
+
+    // If support for Bun should be added to the default "JavaScript Debug Terminal".
+    "bun.debugTerminal.enabled": true,
+    
+    // If the debugger should stop on the first line of the program.
+    "bun.debugTerminal.stopOnEntry": false,
 }
