+++ examples/sgd.ts
@@ -1,5 +1,5 @@
 import { RawTensor, core_ready } from "../index";
-import { parameter_node, source_node } from "../src/node_factory.ts";
+import { parameter_node, source_node } from "../src/tensor_factory.ts";
 import { mul_acc } from "../src/base/raw_tensor_operations.ts";
 
 // if your runtime does not support top-level await,
