+++ src/base/RawTensor.ts
@@ -1,14 +1,20 @@
 import Shape from "./Shape";
 import Strides from "./Strides";
 import core from "./core/build";
-import tensor_to_string, {tensor_info_to_string} from "./to_string";
-import { get_row_major } from "./util";
-import { ordinal_str } from "./util";
+import {tensor_to_string, ordinal_str, tensor_info_to_string} from "./to_string";
+import {flatten, get_strides_row_major, NDArray} from "./util";
 import * as ops from "./raw_tensor_operations.ts";
 
 enum  STRUCT_LAYOUT { DATA, SHAPE, STRIDES, RANK, NELEM, NDATA, OFFSET, SIZE, ISVIEW }
 const STRUCT_SIZE = Object.entries(STRUCT_LAYOUT).length / 2;
 
+/**
+ * This class is acts as an interface between wasm memory and js.
+ * It contains views of the wasm memory locations where the data, shape, strides, etc.
+ * are stored.
+ * Additionally, it contains some metadata operations that don't need to be tracked
+ * in the computation graph.
+ */
 export class RawTensor {
     private readonly view: Int32Array;
     data: Float32Array; // todo: should be private probably
@@ -46,7 +52,7 @@ export class RawTensor {
         if (n > this.rank - 2)
             throw new Error(`Cannot iterate over ${ordinal_str(n)} axis.`);
 
-        const view = RawTensor.view_from(this, n + 1);
+        const view = RawTensor.view_of(this, n + 1);
         const nelem = this.shape.flatten(this.rank - n)[0];
 
         for (let index = 0; index < nelem; index++) {
@@ -55,7 +61,7 @@ export class RawTensor {
             // incrementing the offset, because the user might access
             // the views even after the iteration process is done
             // todo: this leaves us with the problem of de-allocation, however
-            yield RawTensor.view_from(view, 0, offset);
+            yield RawTensor.view_of(view, 0, offset);
         }
     }
 
@@ -79,7 +85,7 @@ export class RawTensor {
     public ones  = () => this.fill(1);
     public free  = () => ops.free(this);
     public clone = () => ops.clone(this);
-    public create_view = (axis = 0, offset = 0) => RawTensor.view_from(this, axis, offset);
+    public create_view = (axis = 0, offset = 0) => RawTensor.view_of(this, axis, offset);
 
     // metadata operations
     public transpose  = (...permutation: number[]) => ops.transpose(this, permutation);
@@ -96,21 +102,26 @@ export class RawTensor {
     // builders
     static scalar = (scalar?: number): RawTensor => RawTensor.create([1], scalar ? [scalar] : undefined);
     static like = (other: RawTensor): RawTensor => RawTensor.create([...other.shape]);
-    static view_from = (a: RawTensor, axis = 0, offset = 0): RawTensor => new RawTensor(core._create_view(a.ptr, axis, offset));
+    static view_of = (a: RawTensor, axis = 0, offset = 0): RawTensor => new RawTensor(core._create_view(a.ptr, axis, offset));
     static create(shape: number[] | Shape, data?: number[]): RawTensor {
         const _shape = [...shape];
         const nelem = _shape.reduce((acc: number, val: number) => acc * val, 1);
-    
+
         if (data !== undefined && data.length !== nelem)
             throw new Error(`Cannot cast array of size ${data.length} into tensor of shape [${shape}]`);
-    
+
         const ptr = core._create_tensor(shape.length, nelem);
         const new_tensor = new RawTensor(ptr);
     
         if (data !== undefined) new_tensor.data.set(data);
         new_tensor.shape.set(shape);
-        new_tensor.strides.set(get_row_major(_shape));
+        new_tensor.strides.set(get_strides_row_major(_shape));
     
         return new_tensor;
+    }
+
+    static from_array(_data: NDArray): RawTensor {
+        const [shape, data] = flatten(_data);
+        return RawTensor.create(shape, data);
     }
 }
