+++ tests/Tensor.test.ts
@@ -1,173 +1,17 @@
 import { describe, expect, test } from "bun:test";
-import { RawTensor } from "../src/base/RawTensor.ts";
 import { core_ready } from "../src/base/Management";
 import Shape from "../src/base/Shape";
 import Strides from "../src/base/Strides";
+import Tensor from "../src/Tensor.ts";
+import {tensor} from "../src/tensor_factory.ts";
 
 describe("tensor creation", async () => {
     await core_ready;
 
-    test("RawTensor.create()", () => {
-        // creates a tensor through RawTensor.create() and ensures that the created
-        // tensor matches the expectation
-        function init_and_validate_tensor(shape: Shape | number[], strides: number[], data?: number[]) {
-            // create tensor
-            const new_tensor = RawTensor.create(shape, data);
-
-            // @ts-expect-error Obscure signature incompatibility between Int32Array.reduce() and Array.reduce()
-            const expected_nelem = shape.reduce((acc: number, val: number) => acc * val, 1);
-
-            // check view types
-            expect(new_tensor).toBeInstanceOf(RawTensor);
-            expect(new_tensor.shape).toBeInstanceOf(Shape);
-            expect(new_tensor.strides).toBeInstanceOf(Strides);
-            expect(new_tensor.data).toBeInstanceOf(Float32Array);
-
-            // ensure that rank and get_nelem() return correct values
-            expect(new_tensor.rank).toBe(shape.length);
-            expect(new_tensor.nelem).toBe(expected_nelem);
-            if (data !== undefined) expect(data.length).toBe(expected_nelem);
-
-            // ensure that the views have the correct sizes
-            expect(new_tensor.shape.length).toBe(shape.length);
-            expect(new_tensor.strides.length).toBe(shape.length);
-            expect(new_tensor.data.length).toBe(expected_nelem);
-
-            // ensure that the shape array is populated with correct data
-            expect([...new_tensor.shape]).toEqual([...shape]);
-            expect([...new_tensor.strides]).toEqual([...strides]);
-            if (data !== undefined) expect([...new_tensor.data]).toEqual([...data]);
-
-            new_tensor.free();
-        }
-
-        // create and validate tensors of various shapes from arrays
-        init_and_validate_tensor([3, 7, 2, 8],          [112, 16, 8, 1]);
-        init_and_validate_tensor([3, 7, 2, 8],          [112, 16, 8, 1],    (new Array(3 * 7 * 2 * 8)).fill(1));
-        init_and_validate_tensor([645],                 [1],                (new Array(645)).fill(1));
-        init_and_validate_tensor([1, 645],              [645, 1],           (new Array(1 * 645)).fill(1));
-        init_and_validate_tensor([1, 78],               [78, 1],            (new Array(1 * 78)).fill(1));
-        init_and_validate_tensor([1, 1, 1, 1, 1, 1],    [1, 1, 1, 1, 1, 1], (new Array(1 * 1 * 1 * 1 * 1 * 1)).fill(1));
-        init_and_validate_tensor([1, 1, 1, 0, 1, 1],    [0, 0, 0, 1, 1, 1], (new Array(1 * 1 * 1 * 0 * 1 * 1)).fill(1));
-
-        const new_shape = new Shape([3, 7, 2, 8]);
-        init_and_validate_tensor(new_shape,          [112, 16, 8, 1]);
-        init_and_validate_tensor(new_shape,          [112, 16, 8, 1],    (new Array(3 * 7 * 2 * 8)).fill(1));
-    });
-
-    // ensures that modifying a tensor does not cause changes in another tensor
-    function check_tensor_independence(old_tensor: RawTensor, new_tensor: RawTensor) {
-        expect(new_tensor.ptr).not.toBe(old_tensor.ptr);
-        expect(new_tensor.shape_ptr).not.toBe(old_tensor.shape_ptr);
-        expect(new_tensor.strides_ptr).not.toBe(old_tensor.strides_ptr);
-        expect(new_tensor.data_ptr).not.toBe(old_tensor.data_ptr);
-
-        // modify new tensor
-        new_tensor.shape[0] = 50;
-        new_tensor.strides[0] = 50;
-        new_tensor.data[0] = 50;
-        // new_tensor.set_rank(50); // todo
-        // new_tensor.set_nelem(50);
-
-        // modify old tensor
-        old_tensor.shape[0] = 10;
-        old_tensor.strides[0] = 10;
-        old_tensor.data[0] = 10;
-        // old_tensor.set_rank(10);
-        // old_tensor.set_nelem(10);
-
-        // check that new tensor was not changed
-        expect([...new_tensor.shape]).not.toEqual([...old_tensor.shape]);
-        expect([...new_tensor.strides]).not.toEqual([...old_tensor.strides]);
-        expect([...new_tensor.data]).not.toEqual([...old_tensor.data]);
-        // expect(new_tensor.rank).not.toEqual(old_tensor.rank);
-        // expect(new_tensor.nelem).not.toEqual(old_tensor.nelem);
-    }
-
-    function check_tensor_metadata_equality(a: RawTensor, b: RawTensor) {
-        expect(a.rank).toBe(b.rank);
-        expect(a.nelem).toBe(b.nelem);
-        expect([...a.shape]).toEqual([...b.shape]);
-        expect([...a.strides]).toEqual([...b.strides]);
-    }
-
-    test("RawTensor.like()", () => {
-        let old_tensor = RawTensor.create([1, 2, 3, 4], (new Array(1 * 2 * 3 * 4)).fill(1));
-        let new_tensor = RawTensor.like(old_tensor);
-
-        check_tensor_metadata_equality(new_tensor, old_tensor);
-        check_tensor_independence(old_tensor, new_tensor);
-
-        old_tensor.free();
-        new_tensor.free();
-
-        // make sure that data was *not* copied to the new tensor        
-        old_tensor = RawTensor.create([1, 2, 3, 4]).rand();
-        new_tensor = RawTensor.like(old_tensor);
-        expect([...old_tensor.data]).not.toEqual([...new_tensor.data]);
-
-        old_tensor.free();
-        new_tensor.free();
-    });
-
-    test("RawTensor.clone()", () => {
-        let old_tensor = RawTensor.create([1, 2, 3, 4], (new Array(1 * 2 * 3 * 4)).fill(1));
-        let new_tensor = old_tensor.clone();
-
-        // make sure metadata is copied correctly and that tensors are independent of one another
-        check_tensor_metadata_equality(new_tensor, old_tensor);
-        check_tensor_independence(old_tensor, new_tensor);
-
-        old_tensor.free();
-        new_tensor.free();
-
-        // make sure that data *was* copied to the new tensor properly      
-        old_tensor = RawTensor.create([1, 2, 3, 4]).rand();
-        new_tensor = old_tensor.clone();
-        expect([...old_tensor.data]).toEqual([...new_tensor.data]);
-
-        old_tensor.free();
-        new_tensor.free();
-    });
-
-    // basic tests for tensor address space overlaps
-    // not really tests for views
-    test("RawTensor views/data-dependence", () => {
-        const old_tensor = RawTensor.create([1, 2, 3, 4]).zeros();
-        const new_tensor = old_tensor.create_view();
-
-        // make sure metadata is copied correctly and that tensors are independent of one another
-        check_tensor_metadata_equality(new_tensor, old_tensor);
-        old_tensor.data[0] = 4;
-        expect(old_tensor.data[0]).toEqual(new_tensor.data[0]);
-    });
-
-    const t1 = RawTensor.create([2, 2, 3], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
-    const t2 = RawTensor.create([3, 2], [1, 2, 3, 4, 5, 6]);
-    const t13 = RawTensor.create([2, 2, 5, 2], [0.261, 0.983, 0.857, 0.279, 0.211, 0.75, 0.671, 0.32, 0.641, 0.317, 0.003, 0.951, 0.332, 0.226, 0.409, 0.475, 0.348, 0.205, 0.11, 0.353, 0.557, 0.309, 0.06, 0.963, 0.368, 0.607, 0.047, 0.52, 0.5, 0.532, 0.138, 0.686, 0.982, 0.134, 0.984, 0.488, 0.856, 0.766, 0.208, 0.29]);
-    const t14 = RawTensor.create([5, 5], [0.93, 0.793, 0.149, 0.827, 0.167, 0.563, 0.485, 0.969, 0.63, 0.909, 0.113, 0.11, 0.627, 0.198, 0.708, 0.574, 0.565, 0.678, 0.013, 0.195, 0.913, 0.131, 0.016, 0.418, 0.277]);
-
-    describe("views", () => {
-        test("get_axis_iterable", () => {
-            // check that subtensors have appropriate shapes
-            for (const e of t1.get_axis_iterable(0)) {
-                expect_shape(e, 2, 3);
-                expect_shape(e.T, 3, 2);
-            }
-
-            for (const e of t1.get_axis_iterable(1)) {
-                expect_shape(e, 3);
-            }
-
-            // todo stride
-
-            // things we should test:
-            // - proper handling of strides of axis_iterable
-            // - proper offsets in get_axis_iterable
-            // - transposition correctness (shape, stride, offset)
-            // - views of views
-        });
-    });
+    const t1 = tensor([2, 2, 3], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]);
+    const t2  = tensor([3, 2], [1, 2, 3, 4, 5, 6]);
+    const t13 = tensor([2, 2, 5, 2], [0.261, 0.983, 0.857, 0.279, 0.211, 0.75, 0.671, 0.32, 0.641, 0.317, 0.003, 0.951, 0.332, 0.226, 0.409, 0.475, 0.348, 0.205, 0.11, 0.353, 0.557, 0.309, 0.06, 0.963, 0.368, 0.607, 0.047, 0.52, 0.5, 0.532, 0.138, 0.686, 0.982, 0.134, 0.984, 0.488, 0.856, 0.766, 0.208, 0.29]);
+    const t14 = tensor([5, 5], [0.93, 0.793, 0.149, 0.827, 0.167, 0.563, 0.485, 0.969, 0.63, 0.909, 0.113, 0.11, 0.627, 0.198, 0.708, 0.574, 0.565, 0.678, 0.013, 0.195, 0.913, 0.131, 0.016, 0.418, 0.277]);
 
     test("transpose", () => {
         // make sure invalid permutations are caught
@@ -346,10 +190,6 @@ describe("tensor creation", async () => {
     });
 });
 
-function expect_shape(t: RawTensor, ...shape: number[]) {
-    expect([...t.shape]).toEqual(shape);
-}
-
 function expect_arrays_closeto(a: number[] | Float32Array, b: number[] | Float32Array) {
     [...a].map((v, i) => {
         if (Number.isNaN(v)) expect(b[i]).toBeNaN();
@@ -358,15 +198,15 @@ function expect_arrays_closeto(a: number[] | Float32Array, b: number[] | Float32
 }
 
 function test_chained_ops(
-    input_tensor:     RawTensor,
-    operations:       (tensor: RawTensor) => RawTensor,
+    input_tensor:     Tensor,
+    operations:       (tensor: Tensor) => Tensor,
     expected_data:    number[] | Float32Array,
     expected_shape:   number[] | Shape,
     expected_strides: number[] | Strides
 ) {
     const result = operations(input_tensor);
 
-    expect([...result.shape]).toEqual([...expected_shape]);
-    expect([...result.strides]).toEqual([...expected_strides]);
-    expect_arrays_closeto(result.data, expected_data);
+    expect([...result.value.shape]).toEqual([...expected_shape]);
+    expect([...result.value.strides]).toEqual([...expected_strides]);
+    expect_arrays_closeto(result.value.data, expected_data);
 }
